Extract code from `prefer-array-index-of.js`
